+++ example/lib/platforms.dart
@@ -1,6 +1,8 @@
 
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:responsive_widgets_prefix/responsive_helper.dart';
+
 
 class PlatformApp extends PlatformWidget {
 
@@ -9,40 +11,46 @@ class PlatformApp extends PlatformWidget {
     required MaterialApp androidApp,
     required CupertinoApp iosApp,
     required TargetPlatform defaultPlatform,
+    required ScreenType defaultScreenType,
   }) : super(key: key,
       androidBuilder: (BuildContext context) => androidApp,
       iosBuilder:  (BuildContext context) => iosApp
     ) {
-      PlatformWidget.setPlatform(defaultPlatform);
+      PlatformWidget.setPlatform(defaultPlatform, defaultScreenType);
   }
 
 }
 
 class PlatformWidget extends StatefulWidget {
   
-  static TargetPlatform? _defaultPlatform;
+  static TargetPlatform? _currentPlatform;
 
   static get platform {
-      if(_defaultPlatform == null) {
+      if(_currentPlatform == null) {
         return TargetPlatform.android;
       }
-      return _defaultPlatform;
+      return _currentPlatform;
+  }
+
+  static get screenType {
+      if(ResponsiveHelper.testingScreenType == ScreenType.None) {
+        return ScreenType.MediumPhone;
+      }
+      return ResponsiveHelper.testingScreenType;
   }
 
   static get isAndroid {
-      return _defaultPlatform == TargetPlatform.android;
+      return _currentPlatform == TargetPlatform.android;
   }
 
   static get isIOS {
-      return _defaultPlatform == TargetPlatform.iOS;
+      return _currentPlatform == TargetPlatform.iOS;
   }
 
-  static void setPlatform(TargetPlatform platform) {
-      _defaultPlatform = platform;
-  }
-
-  static void reassembleApplication() {
-      WidgetsBinding.instance!.reassembleApplication();
+  static void setPlatform(TargetPlatform platform, ScreenType screenType) {
+      _currentPlatform = platform;
+      ResponsiveHelper.testingScreenType = screenType;
+      ResponsiveHelper.updateTestingScreenType();
   }
 
   const PlatformWidget({
@@ -61,13 +69,13 @@ class PlatformWidget extends StatefulWidget {
 class _PlatformWidgetState extends State<PlatformWidget> {
   @override
   Widget build(context) {
-    switch (PlatformWidget._defaultPlatform) {
+    switch (PlatformWidget._currentPlatform) {
       case TargetPlatform.android:
         return widget.androidBuilder(context);
       case TargetPlatform.iOS:      
         return widget.iosBuilder(context);        
       default:
-        assert(false, 'Unexpected platform ${PlatformWidget._defaultPlatform}');
+        assert(false, 'Unexpected platform ${PlatformWidget._currentPlatform}');
         return Container();
     }
   }