+++ lib/integration_test.dart
@@ -4,12 +4,15 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:device_preview_community/device_preview_community.dart';
 import 'package:device_preview_community/src/views/tool_panel/sections/subsections/device_model.dart';
 import 'package:integration_test_helper/integration_test_helper.dart';
+import 'package:provider/provider.dart';
 
 import 'integration_test_binding.dart';
 
 abstract class IntegrationTestPreview extends BaseIntegrationTest {
 
-  late bool _shouldNotShowFrame = true;
+  late bool _shouldShowToolbar = false;
+
+  late bool _shouldNotShowFrame = false;
 
   late bool _testCompleted = false;
 
@@ -25,7 +28,7 @@ abstract class IntegrationTestPreview extends BaseIntegrationTest {
 
   Future<void> testDeviceEndToEnd(DeviceInfo device);
 
-  Future<void> togglePlatformUI(TargetPlatform platform);
+  Future<void> toggleDeviceUI(DeviceInfo device);
 
   Future<BuildContext> getBuildContext();
 
@@ -40,49 +43,54 @@ abstract class IntegrationTestPreview extends BaseIntegrationTest {
     return Future.value(defaultTargetPlatform == TargetPlatform.android);
 
   }
-  
-  @override
-  Future<void> initializeTests(WidgetTester tester, Widget main) async {
 
-        this.tester = tester;
-        WidgetsApp.debugAllowBannerOverride = false;
-        await tester.pumpWidget(DevicePreview(
-            enabled: true,
-            tools: const [
-              ...DevicePreview.defaultTools,
-            ],
-            builder: (context) => main,
-        ));
-
-        await setupInitialData();
-        await waitForUI(durationMultiple: 2);
-
-  }
-
-  Future<void> initializeDevices(List<DeviceInfo> testDevices, {bool hideFrame = false, ScreenshotState state = ScreenshotState.NONE}) async {
+  Future<void> initializeDevices(List<DeviceInfo> testDevices, {bool showToolbar = false, bool hideFrame = false, ScreenshotState state = ScreenshotState.NONE}) async {
+    _shouldShowToolbar = showToolbar;
     _shouldNotShowFrame = hideFrame;
     _screenshotState = state;
     _testOrder.clear();
     _testOrder.addAll(testDevices);
   }
+  
+  @override
+  Future<void> initializeTests(WidgetTester tester, Widget main) async {
+        
+        assert(_testOrder.isNotEmpty, 'There needs to be at least one device to test');
 
-  Future<void> testDevicesEndToEnd() async {
-    
-    assert(_testOrder.isNotEmpty, 'There needs to be at least one device to test');
+        this.tester = tester;
+        WidgetsApp.debugAllowBannerOverride = false;
+        final devicePreview = DevicePreview(
+            enabled: true,
+            defaultDevice: _testOrder[0],
+            isToolbarVisible: _shouldShowToolbar,
+            tools: const [],
+            builder: (context) => main
+        );
+        await tester.pumpWidget(devicePreview).then((value) async {
+            
+            await setupInitialData();
+            await _testDevicesEndToEnd();
+
+        });
+        
+  }
+
+  Future<void> _testDevicesEndToEnd() async {
 
     final testOrderLocalCopy = List.from(_testOrder);
     for (int index = 0; index < testOrderLocalCopy.length; index++) {
       await _startTest();
-      await _togglePlatform();
       await testDeviceEndToEnd(testOrderLocalCopy[index]);
       await _selectNextScreenType();
       await _completedTest();
       await waitForUI();
     }
     await waitForUI(durationMultiple: 3);
+
   }
 
-  Future<void> takeScreenshot(String filePath) async {      
+  @override
+  Future<void> takeScreenshot(String filePath) async {  
       await waitForUI();
       if(_screenshotState == ScreenshotState.PREVIEW) {
           final context = await getBuildContext();
@@ -119,39 +127,46 @@ abstract class IntegrationTestPreview extends BaseIntegrationTest {
 
   }
 
+  Future<void> _updateDevice() async {
+
+    if (! await _hasScreenOption()) return;
+
+    final device = _testOrder[0];
+
+    await toggleDeviceUI(device);
+    WidgetsBinding.instance!.reassembleApplication();
+
+    final isPlatformTypeAndroid = await isPlatformAndroid();
+    final isDevicePlatformAndroid = device.identifier.platform == TargetPlatform.android;
+    final platformTypeIOS = !isPlatformTypeAndroid && !isDevicePlatformAndroid;
+    final platformTypeAndroid = isPlatformTypeAndroid && isDevicePlatformAndroid;
+
+    assert(platformTypeIOS || platformTypeAndroid, 'The platform types do not match. Set the correct app type (Material/Cupertino) for your device (${device.identifier})');
+
+    await waitForUI(durationMultiple: 3).then((value) async {
+      DevicePreview.selectDevice(await getBuildContext(), device.identifier);
+    });
+
+  }
+
+  Future<void> _setPreviewState() async {
+
+    final context = await getBuildContext();
+    DevicePreviewStore state = Provider.of<DevicePreviewStore>(context, listen: false);
+    state.settings = state.settings.copyWith(
+        backgroundTheme: DevicePreviewBackgroundThemeData.dark,
+    );
+    state.notifyListeners();
+
+  }
+
   Future<void> _selectNextScreenType() async {
 
     final device = await _nextScreenOption();
 
     if (! await _hasScreenOption()) return;
 
-    final platformType = device.identifier.platform;
-
-    await tapForKey('model');
-
-    if (TargetPlatform.android == platformType) {
-      await tapWidget('android');
-    } else if (TargetPlatform.iOS == platformType) {
-      await tapWidget('iOS');
-    }
-
-    var newScreenName = device.name;
-    await scrollToListItemText('DeviceListView', newScreenName);
-
-    await tapWidget(newScreenName);
-
-    var popoverClose = find.byTooltip('Back');
-    await tester.tap(popoverClose);
-    await waitForUI();
-
-  }
-
-  Future<void> _togglePlatform() async {
-    if (! await _hasScreenOption()) return;
-
-    final device = _testOrder[0];
-    final platformType = device.identifier.platform;
-    await togglePlatformUI(platformType);
+    await _updateDevice();
 
   }
 
@@ -160,32 +175,18 @@ abstract class IntegrationTestPreview extends BaseIntegrationTest {
       if (_shouldNotShowFrame &&
           await verifyTextForKey('frame-state', 'Visible',
               shouldThrowError: false)) {
-        await tapWidget('Frame visibility');
+        // await tapWidget('Frame visibility');
       }
 
       final device = await _currentScreenOption();
       final platformType = device.identifier.platform;
 
-      await tapForKey('model');
-
-      if (platformType == TargetPlatform.iOS) {
-        await tapWidget('iOS');
-      } else {
-        await tapWidget('android');
-      }
-
       if (defaultTargetPlatform != platformType) {
-        await _togglePlatform();
+        await _updateDevice();
+        await _setPreviewState();
         await waitForUI();
       }
 
-      var newScreenName = device.name;
-
-      await scrollToListItemText('DeviceListView', newScreenName);
-      await tapWidget(newScreenName);
-
-      var popoverClose = find.byTooltip('Back');
-      await tester.tap(popoverClose);
       await waitForUI();
     }
   }
@@ -193,8 +194,6 @@ abstract class IntegrationTestPreview extends BaseIntegrationTest {
   Future<void> _validateLastTestCompleted() async {
     if (!_testCompleted) {
       await _resetTestEnvironment();
-      // if (driver != null) await driver.close();
-      // driver = null;
     }
   }