+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.0.5
+
+* Adding responsive_widgets_prefix to example 
+* Updating screenshots for README
+* Adding assertion that the platform types match
+* Adding call to DevicePreview.selectDevice(...)
+* Adding call to WidgetsBinding.instance!.reassembleApplication()
+* Defaulting to no Device Preview Toolbar
+* Breaking change "Refactored testDevicesEndToEnd to private"
+* Breaking change "Replacing togglePlatformUI \w toggleDeviceUI"
+
 ## 0.0.4
 
 * Upgrade of Integration Test Helper (0.0.6)
@@ -15,4 +26,4 @@
 
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Initial release.